refactor: clean up weather tools in create server
