+++ mcps/mcp-with-express/src/types/geo.types.ts
@@ -1,0 +1,1 @@
+export type State = string;
